Configure VPNKIT IP addresses automatically
WSL2 IP address is not fixed, it can be different after every WSL2 Linux starting. We can get the current VPNKIT IP addresses by "ip route show" and "grep" commands, always.
Further development should be in future that, we need to check VPNKIT_LOWEST_IP and VPNKIT_HIGHEST_IP because they are not allowed to be in conflict with others.